Fix: Calculate tip as percentage, ensure correct division

+++ Day2/Day2 Project.py
@@ -8,5 +8,7 @@ elif tip == 12:
 elif tip == 15:
     tip = 0.15,
 else:
-    tip = float(tip / 100)
+    tip = (tip / 100)
+
+    
 print(f"{tip:.2f}")
